Cleanup of generate-report methods, plus separating out the report generation to new methods and moving reports to the config directory
